Mostly working tests
These step definitions and modified scenarios cover most of what the
resource server needs to do. I haven’t covered performance logs and I
have a niggling feeling the scale/stretch tests won’t actually fail if
it doesn’t work, but this should be enough to get a feel for the
completeness of a resource server!